feat: remove unnecessary components from landing page
